Moved test in common to separated file
